Refactor #2 - details view moves to its own module.
- Move details view to its own module
- Move display_backlink and get_pdf_uri to utils file as they are shared between views
