+++ judgments/urls.py
@@ -2,6 +2,7 @@ from django.urls import path, re_path, register_converter
 
 from . import converters, feeds, views
 from .views.browse import browse
+from .views.detail import detail
 
 register_converter(converters.YearConverter, "yyyy")
 register_converter(converters.DateConverter, "date")
@@ -47,8 +48,8 @@ urlpatterns = [
     re_path(
         r"(?P<judgment_uri>.*/\d{4}/\d+)/data.xml", views.detail_xml, name="detail_xml"
     ),
-    re_path(r"(?P<judgment_uri>.*/\d{4}/\d+)/data.html", views.detail, name="detail"),
-    re_path(r"(?P<judgment_uri>.*/\d{4}/\d+)", views.detail, name="detail"),
+    re_path(r"(?P<judgment_uri>.*/\d{4}/\d+)/data.html", detail, name="detail"),
+    re_path(r"(?P<judgment_uri>.*/\d{4}/\d+)", detail, name="detail"),
     path("judgments/results", views.results, name="results"),
     path("judgments/advanced_search", views.advanced_search, name="advanced_search"),
     path("", views.index, name="home"),
