+++ judgments/views/__init__.py
@@ -2,9 +2,7 @@ import logging
 import os
 import re
 import urllib
-from urllib.parse import urlparse
 
-import environ
 import requests
 from caselawclient.Client import (
     RESULTS_PER_PAGE,
@@ -16,7 +14,6 @@ from django.conf import settings
 from django.http import Http404, HttpResponse
 from django.shortcuts import redirect
 from django.template import loader
-from django.template.defaultfilters import filesizeformat
 from django.template.response import TemplateResponse
 from django.urls import reverse
 from django.utils.translation import gettext
@@ -30,36 +27,10 @@ from judgments.models import SearchResult
 from judgments.utils import (
     MAX_RESULTS_PER_PAGE,
     as_integer,
+    get_pdf_uri,
     paginator,
     perform_advanced_search,
 )
-
-env = environ.Env()
-
-
-def detail(request, judgment_uri):
-    context = {}
-    try:
-        is_published = api_client.get_published(judgment_uri)
-    except MarklogicAPIError:
-        raise Http404("Judgment was not found")
-
-    if is_published:
-        try:
-            results = api_client.eval_xslt(judgment_uri)
-            multipart_data = decoder.MultipartDecoder.from_response(results)
-            judgment = multipart_data.parts[0].text
-            context["judgment"] = judgment
-            context["page_title"] = api_client.get_judgment_name(judgment_uri)
-            context["judgment_uri"] = judgment_uri
-            context["pdf_size"] = get_pdf_size(judgment_uri)
-            context["back_link"] = get_back_link(request)
-        except MarklogicResourceNotFoundError:
-            raise Http404("Judgment was not found")
-        template = loader.get_template("judgment/detail.html")
-        return TemplateResponse(request, template, context={"context": context})
-    else:
-        raise Http404("This Judgment is not available")
 
 
 def advanced_search(request):
@@ -161,22 +132,6 @@ class PdfDetailView(WeasyTemplateResponseMixin, TemplateView):
         context["judgment"] = multipart_data.parts[0].text
 
         return context
-
-
-def get_pdf_uri(judgment_uri):
-    """Create a string saying where the S3 PDF will be for a judgment uri"""
-    pdf_path = f'{judgment_uri}/{judgment_uri.replace("/", "_")}.pdf'
-    return f'https://{env("PUBLIC_ASSET_BUCKET")}.s3.{env("S3_REGION")}.amazonaws.com/{pdf_path}'
-
-
-def get_pdf_size(judgment_uri):
-    """Return the size of the S3 PDF for a judgment as a string in brackets, or an empty string if unavailable"""
-    response = requests.head(get_pdf_uri(judgment_uri))
-    content_length = response.headers.get("Content-Length", None)
-    if content_length:
-        filesize = filesizeformat(int(content_length))
-        return f" ({filesize})"
-    return ""
 
 
 def get_best_pdf(request, judgment_uri):
@@ -282,22 +237,6 @@ def results(request):
 
 def trim_leading_slash(uri):
     return re.sub("^/|/$", "", uri)
-
-
-def get_back_link(request):
-    back_link = request.META.get("HTTP_REFERER")
-    if display_back_link(back_link):
-        return back_link
-    else:
-        return None
-
-
-def display_back_link(back_link):
-    if back_link:
-        url = urlparse(back_link)
-        return url.path in ["/judgments/results", "/judgments/advanced_search"]
-    else:
-        return False
 
 
 def has_filters(query_params, exclude=["order", "per_page"]):
